Split up concerns in react
