+++ src/schedulion-react-fe/src/components/home/Home.css
@@ -1,5 +1,5 @@
 .Home {
-  background-image: url(./images/hank.png);
+  background-image: url(../images/hank.png);
   position: fixed;
   min-width: 100%;
   min-height: 100%;
