+++ src/schedulion-react-fe/src/components/landing_page/LandingPage.css
@@ -1,5 +1,5 @@
 .LandingPage {
-  background-image: url(./images/landing.png);
+  background-image: url(../images/landing.png);
   position: fixed;
   min-width: 100%;
   min-height: 100%;
